Do the heaviest test first. Another massive speedup.
We can always rearrange the columns to make our first test conform
to the "animals are introduced left-to-right" rule. So, without loss
of generality, we can let our first test (test number 1) be the one
involving the most sheep, and then each subsequent test must use
the same number of animals or fewer.

This greatly speeds up the unproductive early searches.

    `time ./st 11 2 8` before:  3.067 seconds
    `time ./st 11 2 8` after:   0.027 seconds

    `time ./st 12 2 8` before:  >10 minutes
    `time ./st 12 2 8` after:   0.107 seconds