Use a better test-ordering filter.
Rather than saying that each test must introduce a certain number
of _new_ animals (which quickly devolves to introducing 1 or even 0
animals at a time), let's say that each test must involve no more
animals (irrespective of newness) than the previous test did.

With the previous filter, we would happily pursue
    T T T . . . . . .
    . . . T T T . . .
    . . . . . . T . .
all the way to the end of a very large search space.
With the new filter, that search space is very small, because
each subsequent test cannot have more than 1 animal in it.
(And the space quickly narrows to zero, because each test's numerical
value must be greater than the preceding test, as well.)

This is useful when the answer is "no solution exists."

        `time ./st 13 2 8` before:  39.645 seconds
        `time ./st 13 2 8` after:   11.175 seconds