Update the README with OEIS sequences.
Some of the data in the README's triangle now comes from OEIS-hearsay,
not from the programs in this repository.

Add a version of `choose<n,k>()` that doesn't exceed compilation limits,
so that I can stop computing those values by hand.

Add a sample solution for t(273,5)=91, but that's no help toward
any better solution for t(100,5).