TangentCircles/find-edwardh.py: Prune more of the search space
The seven thetas we find must divide the circle into seven pieces,
of which the first is the biggest; by the pigeonhole principle
that must be at least 1/7 of the whole 2pi radians (so we needn't
search into any `t1` that isn't at least that big). `t7` could be
up to epsilon-less-than `t1`; but then `t2` must account for at least
1/5 of what remains. `t3` must account for at least 1/4 of what remains
after subtracting `2*t1 + t2`. `t4` must account for at least 1/3
of what remains after that. And so on.

This drastically prunes the number of combinations we bother
looking at. `python find-edwardh.py 100` now gives its first four
results in under 1 second on my machine.