PolybarPolyominoes: speeding it up
Shrinking the length of the strings shrinks the depth of the trie,
and utilizes more of the child-pointer-buckets in each trie node.
This greatly improves memory efficiency, which might help with
counting decadominoes. I see it by default (without `-DUSE_TSL_HTRIE`)
using only like 10% CPU, indicating that it's spending most of its
time chasing pointers in memory (which means in swap space).
This should help with that.